Optimization round: median 1.04x on DaCapo (paper target was 1.06x)
Site-level VERSION_GATE checks at every emitted pre-hook collapse the
"agent attached, no checkpoint ever called" steady-state cost to a
single volatile-int read + branch that C2 speculatively eliminates via
uncommon trap. Combined with skipping immutable JDK leaves (String, the
Number boxed-primitive hierarchy), this carries the median DaCapo
ratio from the post-correctness 1.52x down to 1.04x — at-or-below the
2018 paper's 1.06x geomean — while keeping all correctness gates green
(20/20 demos, 35/35 unit tests, 22/22 DaCapo functional, 320/320 paper
§5.1 microbench).

Three primary changes:

1. Site-level VERSION_GATE emit (FieldAccessWrapper, StaticFieldRewriter,
   ArrayAccessWrapper). Previously the wrap site emitted a direct
   helper call (INVOKESTATIC fieldAccess / noteStaticAccess /
   beforeStore) whose body checked the gate. Now the wrap site emits
   GETSTATIC VERSION_GATE; IFEQ skip; <hook>; skip: directly. Lets
   interpreter and C1 tier skip the dispatch entirely, and lets C2
   constant-fold the dead branch when the gate is profile-zero.
   Recovers 1.52x → 1.10x median.

2. Skip java.lang.String. Final, immutable, extends Object directly
   (no inheritance trap). Removes 12 bytes per String instance and
   strips wraps from heavily-called String.hashCode/equals/charAt.
   Recovers 1.10x → 1.06x.

3. Skip java.lang.Number boxed-primitive hierarchy (Number itself plus
   Integer/Long/Float/Double/Boolean/Short/Character; Byte already
   skipped). Number must be skipped together with the leaves: with
   only the leaves skipped, Integer instances inherit Number's
   $$crochetCheckpoint which calls allocateShadow(Number.class) and
   fails because Number is abstract. With both skipped, Integer carries
   no $$crochet surface; the instanceof CRIJInstrumented check at every
   propagation site returns false on Integer, so we walk past without
   invoking. BigInteger/BigDecimal/Atomic* (also Number subclasses but
   concrete and mutable) stay instrumented; their emit checks
   superIsInstrumented(Number) → false and skips the super-chain call
   cleanly. Recovers 1.06x → 1.04x.

Supporting:

- VERSION_GATE switched from volatile long to volatile int (saves
  LCONST_0 + LCMP per emit site).
- Removed redundant READY check from gate logic — VERSION_GATE != 0
  causally implies READY == true (checkpoint can only fire after
  premain finishes).
- interceptedArraycopy gated similarly — saves the ArrayRegistry
  metaFor lookup per intercepted copy.
- ArrayRegistry.warmup() invoked from CrochetAgent.install to preload
  the inner-class closure (ProbeKey, IdKey, ArrayMeta). Required because
  the gate now skips the bootstrap-time arraycopy chain that previously
  eagerly loaded ArrayRegistry; the first non-zero-gate call would
  otherwise land mid-TransformerWrapper.transform and trip
  ClassCircularityError.

Per-bench v1 → v6 highlights (median of 6 measurements across 2 sweeps):

  biojava    5.60x → 1.00x   (-4.60)
  lusearch   3.23x → 0.98x   (-2.25)
  tradebeans 3.23x → 1.14x   (-2.09)
  eclipse    2.52x → 0.98x   (-1.54)
  graphchi   2.34x → 1.14x   (-1.20)
  h2         2.28x → 1.10x   (-1.18)
  tradesoap  2.24x → 1.15x   (-1.09)
  jython     1.95x → 1.01x   (-0.94)
  pmd        1.55x → 0.98x   (-0.57)
  spring     1.48x → 1.00x   (-0.48)

8 of 22 benches now run at-or-below baseline. Sunflow (1.36x) is the
worst remaining — CPU-heavy ray tracing where the inlined GETSTATIC +
IFEQ per user-class field access is the floor we can't push lower
without per-method skip heuristics.

Detailed methodology, per-benchmark numbers, and architectural
discussion in BENCHMARK.md §10.

+++ crochet-agent/src/main/java/net/jonbell/crochet/runtime/VersionCounter.java
@@ -73,7 +73,7 @@ final class VersionCounter {
                 // inside RuntimeReady without touching any other runtime
                 // class. See {@code RuntimeReady#VERSION_GATE} for why
                 // this gate is load-bearing during JVM bootstrap.
-                RuntimeReady.VERSION_GATE = next;
+                RuntimeReady.VERSION_GATE = (int) next;
                 // Gap 7 closure: a successful first checkpoint implies the
                 // agent runtime is fully loaded. Flip the RuntimeReady gate
                 // so instrumented JDK bytecode's pre-hooks start tracking
@@ -104,7 +104,7 @@ final class VersionCounter {
                 // inside RuntimeReady without touching any other runtime
                 // class. See {@code RuntimeReady#VERSION_GATE} for why
                 // this gate is load-bearing during JVM bootstrap.
-                RuntimeReady.VERSION_GATE = next;
+                RuntimeReady.VERSION_GATE = (int) next;
                 RuntimeReady.markReady();
                 return (int) next;
             }
